hiện tt tk user, thêm sửa xóa địa chỉ user

+++ database/migrations/2024_10_17_142422_update_users_table.php
@@ -1,5 +1,4 @@
-<?php
-
+<?php 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -11,13 +10,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->foreignId('role_id')->constrained('roles')->onDelete('cascade')->after('id');
-        });
+        // Kiểm tra bảng users
+        if (!Schema::hasColumn('users', 'role_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->foreignId('role_id')->constrained('roles')->onDelete('cascade')->after('id');
+            });
+        }
 
-        Schema::table('shipers', function (Blueprint $table) {
-            $table->foreignId('role_id')->constrained('roles')->onDelete('cascade')->after('id');
-        });
+        // Kiểm tra bảng shipers
+        if (!Schema::hasColumn('shipers', 'role_id')) {
+            Schema::table('shipers', function (Blueprint $table) {
+                $table->foreignId('role_id')->constrained('roles')->onDelete('cascade')->after('id');
+            });
+        }
     }
 
     /**
